Rewrite TestEventFunctions in Groovy
This way we can use BlockingQueue.put() without adding noise for exception handling.
